backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["theochitayat@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
psycopg2 = "^2.9.9"
shapely = "^2.0.6"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
